+++ modules/prestaneo/controllers/admin/AdminSyncImportsController.php
@@ -52,7 +52,7 @@ class AdminSyncImportsController extends ModuleAdminController
             'import'        => $this->module->readImportFolder(),
             'link'          => $this->context->link,
             'path'          => '/modules/'.$this->module->name,
-            'cronpath'      => $this->context->shop->getBaseURL().'modules/'.$this->module->name.'/cron.php',
+            'cronpath'      => $this->context->shop->getBaseURL(true).'modules/'.$this->module->name.'/cron.php',
             'mod_sync'      => MOD_SYNC,
             'mod_sync_name' => MOD_SYNC_NAME,
             'languages'     => Language::getLanguages(),